test(import): move AutoEq repository tests to the import target
The catalogue parser and bounded download tests lived in the IPC suite and
were reached through a temporary re-export of GlassEQProfileImport from the
settings UI. They now live next to the code they cover and the re-export is
gone.
