+++ Sources/GlassEQSettingsUI/ProfileImport/ProfileImportSheet.swift
@@ -1,7 +1,7 @@
 import GlassEQCore
 // Re-exported so GlassEQSettingsUI clients keep seeing the import types that
 // used to live in this module.
-@_exported import GlassEQProfileImport
+import GlassEQProfileImport
 import GlassEQSettingsIPC
 import SwiftUI
